Extract share button logic into own component
