+++ frontend/src/routes/Video.tsx
@@ -1,5 +1,4 @@
 import React, {
-    ReactElement,
     ReactNode,
     useEffect,
     useRef,
@@ -12,12 +11,11 @@ import {
 import { useTranslation } from "react-i18next";
 import { fetchQuery, OperationType } from "relay-runtime";
 import {
-    LuCode, LuDownload, LuInfo, LuLink, LuQrCode, LuRss, LuSettings, LuShare2, LuLockOpen,
+    LuCode, LuDownload, LuInfo, LuLink, LuRss, LuSettings, LuLockOpen,
 } from "react-icons/lu";
-import { QRCodeCanvas } from "qrcode.react";
 import {
     match, unreachable, screenWidthAtMost, screenWidthAbove, useColorScheme,
-    Floating, FloatingContainer, FloatingTrigger, WithTooltip, Card, Button, ProtoButton,
+    Floating, FloatingContainer, FloatingTrigger, WithTooltip, Card, Button,
     notNullish,
 } from "@opencast/appkit";
 import { VideoObject, WithContext } from "schema-dts";
@@ -39,7 +37,6 @@ import {
     toIsoDuration,
     useForceRerender,
     translatedConfig,
-    currentRef,
     secondsToTimeString,
     eventId,
     keyOfId,
@@ -74,12 +71,11 @@ import {
 } from "../ui/Blocks/__generated__/PlaylistBlockPlaylistData.graphql";
 import { getEventTimeInfo } from "../util/video";
 import { formatDuration } from "../ui/Video";
-import { ellipsisOverflowCss, focusStyle } from "../ui";
+import { ellipsisOverflowCss } from "../ui";
 import { realmBreadcrumbs } from "../util/realm";
 import { TrackInfo } from "./manage/Video/TechnicalDetails";
 import { COLORS } from "../color";
 import { RelativeDate } from "../ui/time";
-import { Modal, ModalHandle } from "../ui/Modal";
 import { PlayerContextProvider, usePlayerContext } from "../ui/player/PlayerContext";
 import { CollapsibleDescription } from "../ui/metadata";
 import { DirectSeriesRoute, SeriesRoute } from "./Series";
@@ -94,6 +90,7 @@ import { AuthorizedBlockEvent } from "../ui/Blocks/Video";
 import {
     VideoPageAuthorizedData$data, VideoPageAuthorizedData$key,
 } from "./__generated__/VideoPageAuthorizedData.graphql";
+import { QrCodeButton, ShareButton } from "../ui/ShareButton";
 
 
 // ===========================================================================================
@@ -904,7 +901,7 @@ const Metadata: React.FC<MetadataProps> = ({ event, realmPath }) => {
                 {CONFIG.showDownloadButton && event.authorizedData && (
                     <DownloadButton event={event} />
                 )}
-                <ShareButton {...{ event }} />
+                <VideoShareButton {...{ event }} />
             </div>
         </div>
         <div css={{
@@ -973,260 +970,125 @@ const DownloadButton: React.FC<{ event: SyncedEvent }> = ({ event }) => {
     );
 };
 
-const ShareButton: React.FC<{ event: SyncedEvent }> = ({ event }) => {
-    type MenuState = "closed" | "main" | "embed" | "rss";
-    /* eslint-disable react/jsx-key */
-    const entries: [Exclude<MenuState, "closed">, ReactElement][] = [
-        ["main", <LuLink />],
-        ["embed", <LuCode />],
-    ];
-    if (event.series) {
-        entries.push(["rss", <LuRss />]);
-    }
-    /* eslint-enable react/jsx-key */
-
+const VideoShareButton: React.FC<{ event: SyncedEvent }> = ({ event }) => {
     const { t } = useTranslation();
-    const [menuState, setMenuState] = useState<MenuState>("closed");
     const [timestamp, setTimestamp] = useState(0);
     const [addLinkTimestamp, setAddLinkTimestamp] = useState(false);
     const [addEmbedTimestamp, setAddEmbedTimestamp] = useState(false);
-    const isDark = useColorScheme().scheme === "dark";
-    const ref = useRef(null);
-    const qrModalRef = useRef<ModalHandle>(null);
     const { paella, playerIsLoaded } = usePlayerContext();
 
     const timeStringPattern = /\?t=(\d+h)?(\d+m)?(\d+s)?/;
 
-    const isActive = (label: MenuState) => label === menuState;
+    const series = event.series;
+    const tabs = {
+        "main": {
+            label: t("video.share.main"),
+            Icon: LuLink,
+            Component: () => {
+                let url = window.location.href.replace(timeStringPattern, "");
+                url += addLinkTimestamp && timestamp
+                    ? `?t=${secondsToTimeString(timestamp)}`
+                    : "";
 
-    const tabStyle = {
-        display: "flex",
-        flexDirection: "column",
-        flex: `1 calc(100% / ${entries.length})`,
-        backgroundColor: COLORS.neutral20,
-        paddingBottom: 4,
-        cursor: "pointer",
-        alignItems: "center",
-        borderRight: `1px solid ${COLORS.neutral40}`,
-        borderTop: "none",
-        borderBottom: `1px solid ${COLORS.neutral40}`,
-        ":is(:first-child)": { borderTopLeftRadius: 4 },
-        ":is(:last-child)": {
-            borderRight: "none",
-            borderTopRightRadius: 4,
-        },
-        "& > svg": {
-            width: 32,
-            height: 32,
-            color: COLORS.primary1,
-            padding: "8px 4px 4px",
-        },
-        "&[disabled]": {
-            cursor: "default",
-            backgroundColor: isDark ? COLORS.neutral15 : COLORS.neutral05,
-            borderBottom: "none",
-            svg: { color: COLORS.primary0 },
-        },
-        ":not([disabled])": {
-            "&:hover": { backgroundColor: COLORS.neutral15 },
-        },
-        ...focusStyle({ inset: true }),
-
-        // By using the `has()` selector, these styles only get applied
-        // to non-firefox browsers. Once firefox supports that selector,
-        // this border radius stuff should get refactored.
-        ":has(svg)": {
-            "&[disabled]": {
-                borderRight: "none",
-                "+ button": {
-                    borderLeft: `1px solid ${COLORS.neutral40}`,
-                    borderBottomLeftRadius: 4,
-                },
-            },
-            ":not([disabled]):has(+ button[disabled])": {
-                borderBottomRightRadius: 4,
-                borderLeft: "none",
-            },
-        },
-    } as const;
-
-    const header = <div css={{ display: "flex" }}>
-        {entries.map(([label, icon]) => (
-            <ProtoButton
-                disabled={isActive(label)}
-                key={label}
-                onClick={() => setMenuState(label)}
-                css={tabStyle}
-            >
-                {icon}
-                {t(`video.share.${label}`)}
-            </ProtoButton>
-        ))}
-    </div>;
-
-    const ShowQRCodeButton: React.FC<{ target: string; label: MenuState }> = (
-        { target, label },
-    ) => <>
-        <Button
-            onClick={() => currentRef(qrModalRef).open()}
-            css={{ width: "max-content" }}
-        >
-            <LuQrCode />
-            {t("video.share.show-qr-code")}
-        </Button>
-        <Modal
-            ref={qrModalRef}
-            title={t("video.share.title", { title: label })}
-            css={{ minWidth: "max-content" }}
-            closeOnOutsideClick
-        >
-            <div css={{ display: "flex", justifyContent: "center" }}>
-                <QRCodeCanvas
-                    value={target}
-                    size={250}
-                    css={{
-                        margin: 16,
-                        outline: "8px solid #FFFFFF",
-                    }}
-                />
-            </div>
-        </Modal>
-    </>;
-
-    const inner = match(menuState, {
-        "closed": () => null,
-        "main": () => {
-            let url = window.location.href.replace(timeStringPattern, "");
-            url += addLinkTimestamp && timestamp
-                ? `?t=${secondsToTimeString(timestamp)}`
-                : "";
-
-            return <>
-                <div>
-                    <CopyableInput
-                        label={t("manage.my-videos.details.copy-direct-link-to-clipboard")}
-                        css={{ fontSize: 14, width: 400, marginBottom: 6 }}
-                        value={url}
-                    />
-                    {!event.isLive && <InputWithCheckbox
-                        checkboxChecked={addLinkTimestamp}
-                        setCheckboxChecked={setAddLinkTimestamp}
-                        label={t("manage.my-videos.details.set-time")}
-                        input={<TimeInput
-                            {...{ timestamp, setTimestamp }}
-                            disabled={!addLinkTimestamp}
-                        />}
-                    />}
-                </div>
-                <ShowQRCodeButton target={url} label={menuState} />
-            </>;
-        },
-        "embed": () => {
-            const ar = event.authorizedData == null
-                ? [16, 9]
-                : getPlayerAspectRatio(event.authorizedData.tracks);
-
-            const url = new URL(location.href.replace(timeStringPattern, ""));
-            url.search = addEmbedTimestamp && timestamp
-                ? `?t=${secondsToTimeString(timestamp)}`
-                : "";
-            url.pathname = EmbedVideoRoute.url({ videoId: event.id });
-
-            const embedCode = `<iframe ${[
-                'name="Tobira Player"',
-                `src="${url}"`,
-                "allow=fullscreen",
-                `style="${[
-                    "border: none;",
-                    "width: 100%;",
-                    `aspect-ratio: ${ar.join("/")};`,
-                ].join(" ")}"`,
-            ].join(" ")}></iframe>`;
-
-            return <>
-                <div>
-                    <CopyableInput
-                        label={t("video.embed.copy-embed-code-to-clipboard")}
-                        value={embedCode}
-                        multiline
-                        css={{ fontSize: 14, width: 400, height: 75, marginBottom: 6 }}
-                    />
-                    {!event.isLive && <InputWithCheckbox
-                        checkboxChecked={addEmbedTimestamp}
-                        setCheckboxChecked={setAddEmbedTimestamp}
-                        label={t("manage.my-videos.details.set-time")}
-                        input={<TimeInput
-                            {...{ timestamp, setTimestamp }}
-                            disabled={!addEmbedTimestamp}
-                        />}
-                    />}
-                </div>
-                <ShowQRCodeButton target={embedCode} label={menuState} />
-            </>;
-        },
-        "rss": () => {
-            if (event.series) {
-                const rssUrl = window.location.origin + `/~rss/series/${keyOfId(event.series.id)}`;
                 return <>
                     <div>
                         <CopyableInput
-                            label={t("video.rss.copy-link-to-clipboard")}
+                            label={t("manage.my-videos.details.copy-direct-link-to-clipboard")}
                             css={{ fontSize: 14, width: 400, marginBottom: 6 }}
-                            value={rssUrl}
+                            value={url}
                         />
+                        {!event.isLive && <InputWithCheckbox
+                            checkboxChecked={addLinkTimestamp}
+                            setCheckboxChecked={setAddLinkTimestamp}
+                            label={t("manage.my-videos.details.set-time")}
+                            input={<TimeInput
+                                {...{ timestamp, setTimestamp }}
+                                disabled={!addLinkTimestamp}
+                            />}
+                        />}
                     </div>
-                    <ShowQRCodeButton target={rssUrl} label={menuState} />
+                    <QrCodeButton target={url} label={t("video.share.main")} />
                 </>;
-            } else {
-                return null;
-            }
+            },
         },
-    });
+        "embed": {
+            label: t("video.share.embed"),
+            Icon: LuCode,
+            Component: () => {
+                const ar = event.authorizedData == null
+                    ? [16, 9]
+                    : getPlayerAspectRatio(event.authorizedData.tracks);
 
+                const url = new URL(location.href.replace(timeStringPattern, ""));
+                url.search = addEmbedTimestamp && timestamp
+                    ? `?t=${secondsToTimeString(timestamp)}`
+                    : "";
+                url.pathname = EmbedVideoRoute.url({ videoId: event.id });
 
-    return (
-        <FloatingContainer
-            ref={ref}
-            placement="top"
-            arrowSize={12}
-            ariaRole="dialog"
-            open={menuState !== "closed"}
-            onClose={() => setMenuState("closed")}
-            viewPortMargin={12}
-        >
-            <FloatingTrigger>
-                <Button onClick={() => {
-                    setMenuState(state => state === "closed" ? "main" : "closed");
-                    if (playerIsLoaded) {
-                        paella.current?.player.videoContainer.currentTime().then(res => {
-                            setTimestamp(res);
-                        });
-                    }
-                }}>
-                    <LuShare2 size={16} />
-                    {t("general.action.share")}
-                </Button>
-            </FloatingTrigger>
-            <Floating
-                padding={0}
-                backgroundColor={isDark ? COLORS.neutral15 : COLORS.neutral05}
-                css={{
-                    height: 240,
-                    display: "flex",
-                    flexDirection: "column",
-                }}
-            >
-                {header}
-                <div css={{
-                    margin: 16,
-                    flexGrow: 1,
-                    display: "flex",
-                    flexDirection: "column",
-                    justifyContent: "space-between",
-                }}>{inner}</div>
-            </Floating>
-        </FloatingContainer>
-    );
+                const embedCode = `<iframe ${[
+                    'name="Tobira Player"',
+                    `src="${url}"`,
+                    "allow=fullscreen",
+                    `style="${[
+                        "border: none;",
+                        "width: 100%;",
+                        `aspect-ratio: ${ar.join("/")};`,
+                    ].join(" ")}"`,
+                ].join(" ")}></iframe>`;
+
+                return <>
+                    <div>
+                        <CopyableInput
+                            label={t("video.embed.copy-embed-code-to-clipboard")}
+                            value={embedCode}
+                            multiline
+                            css={{ fontSize: 14, width: 400, height: 75, marginBottom: 6 }}
+                        />
+                        {!event.isLive && <InputWithCheckbox
+                            checkboxChecked={addEmbedTimestamp}
+                            setCheckboxChecked={setAddEmbedTimestamp}
+                            label={t("manage.my-videos.details.set-time")}
+                            input={<TimeInput
+                                {...{ timestamp, setTimestamp }}
+                                disabled={!addEmbedTimestamp}
+                            />}
+                        />}
+                    </div>
+                    <QrCodeButton target={embedCode} label={t("video.share.embed")} />
+                </>;
+            },
+        },
+        ...series && {
+            "rss": {
+                label: t("video.share.rss"),
+                Icon: LuRss,
+                Component: () => {
+                    const rssUrl = window.location.origin
+                        + `/~rss/series/${keyOfId(series.id)}`;
+                    return <>
+                        <div>
+                            <CopyableInput
+                                label={t("video.rss.copy-link-to-clipboard")}
+                                css={{ fontSize: 14, width: 400, marginBottom: 6 }}
+                                value={rssUrl}
+                            />
+                        </div>
+                        <QrCodeButton target={rssUrl} label={t("video.share.rss")} />
+                    </>;
+                },
+            },
+        },
+    };
+
+    const onOpen = () => {
+        if (playerIsLoaded) {
+            paella.current?.player.videoContainer.currentTime().then(res => {
+                setTimestamp(res);
+            });
+        }
+    };
+
+    return <ShareButton {...{ tabs, onOpen }} />;
 };
